requirements.txt
opencv-python
scipy
Pillow
matplotlib